requirements.txt
streamlit
openai
python-dotenv
pandas
requests